requirements.txt
django==1.8.4
django-user-management==13.1.1
dj-database-url==0.3.0
flake8==2.4.1
flake8-import-order==0.5.1
pep8==1.6.2
psycopg2==2.6.1
